template_storage/requirements.txt
google-cloud-storage>=1.0.0
google-cloud-vision>=1.0.0
google-api-core>=2.0.0
grpcio>=1.38.0